certificados itau e bradesco

+++ src/Microled.Pix.Domain/Request/PagamentoRequest.cs
@@ -6,7 +6,7 @@ namespace Microled.Pix.Domain.Request
     public class PagamentoRequest : Notifiable<Notification>
     {
         public PagamentoRequest(int id_Empresa, int id_Processo, string numero_Titulo, string devedor_Nome, string devedor_CPF, 
-            string devedor_CNPJ, double valor, string data_Hora_Expiracao_Pagamento, string solicitacao_Pagador)
+            string devedor_CNPJ, decimal valor, string data_Hora_Expiracao_Pagamento, string solicitacao_Pagador)
         {
             Id_Empresa = id_Empresa;
             Id_Processo = id_Processo;
@@ -30,7 +30,7 @@ namespace Microled.Pix.Domain.Request
         public string Devedor_Nome { get; set; }
         public string Devedor_CPF { get; set; }
         public string Devedor_CNPJ { get; set; }
-        public double Valor { get; set; }
+        public decimal Valor { get; set; }
         public string Data_Hora_Expiracao_Pagamento { get; set; } 
         public string Solicitacao_Pagador { get; set; } 
         //public string[] Informacoes_Adicionais { get; set; }
